feature/82 refactor: rebase 이후 기능 동기화
 - 초기("/") page를 지도화면 으로 수정
