+++ front/src/components/post/PostDetailPage.vue
@@ -68,7 +68,7 @@ export default {
   methods: {
     hasLike(like) {
       return (
-        like.memberId === this.$store.getters["member/getMembers"].id &&
+        like.memberId === this.$store.getters["member/getMember"].id &&
         like.postId === this.post.id
       );
     },
